Move controllers into api and routes
