+++ server/routes/controllers/editor-directory-structure-constructor.ts
@@ -1,7 +1,7 @@
 import fs from "fs";
-import customPages from "../config/admin/admin-pages.json";
-import { DirectoryEntry, PageInflow } from "../types/routing";
-import { getFormattedURL } from "../util/util";
+import customPages from "../../config/admin/admin-pages.json";
+import { DirectoryEntry, PageInflow } from "../../types/routing";
+import { getFormattedURL } from "../../util/util";
 
 /**
  * Constructs the /admin/editor menu structure
